Separate ResultConcTable and ResultConcPlot
