+++ frontend/src/pages/Results.tsx
@@ -1,11 +1,12 @@
 import React, { useEffect } from "react";
 import { Tabs } from "@equinor/eds-core-react";
 import { useState } from "react";
-import OutputConcentrations from "./OutputConcentrations";
+import ResultConcPlot from "../components/ConcResultPlot";
 import Reactions from "./Reactions";
 import { SimulationResults } from "../dto/SimulationResults";
 import { useParams } from "react-router-dom";
 import { getSimulationResults } from "../api/api";
+import ResultConcTable from "../components/ConcResultTable";
 
 interface ResultsProps {
     simulationResults?: SimulationResults;
@@ -54,7 +55,8 @@ const Results: React.FC<ResultsProps> = ({ simulationResults }) => {
                 </Tabs.List>
                 <Tabs.Panels>
                     <Tabs.Panel>
-                        <OutputConcentrations simulationResults={results!} />
+                        <ResultConcPlot simulationResults={results!} />
+                        <ResultConcTable initFinalDiff={results!.results.initfinaldiff} />
                     </Tabs.Panel>
                     <Tabs.Panel>
                         <Reactions simulationResults={results!} />
